Workaround for #5

+++ cad_source/components/zmath/uzegeometrytypes.pas
@@ -103,7 +103,7 @@ type
   end;
 {EXPORT+}
 
-  FontFloat=Single;
+  FontFloat=Double;
   PFontFloat=^FontFloat;
   PGDBXCoordinate=^GDBXCoordinate;
   GDBXCoordinate=Double;
